Fix assertion failure
  ERROR: LoadError: AssertionError: fb.this_shot_rotation == -1

introduced by 6725eb3914a1520dc5fd8cf9db97c2176f35dcc1.

Working with a warped single card in a K'nex frame, this is what I think the results should be:

ABCD    +1 rotation
DCBA    -1 rotation

If we start with edge 1 on top, which edge is on top after 1 Forward rotation?

           FrontToTheLeft  FrontToTheRight
                    |          |
                    |          |
BackToFront -----   2          4
FrontToBack -----   2          4

edge 1 -> 4 is ABCD  which is   +1
edge 1 -> 2 is DBCA  which is   -1

I fixed the assertions to match the above.